requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
binance==0.3
bitarray==2.6.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cytoolz==0.12.0
dateparser==1.1.1
Django==3.2.16
django-mailer==2.2
django-online-users==0.3
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lockfile==0.12.2
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
netaddr==0.8.0
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parsimonious==0.8.1
pathspec==0.10.1
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.19.5
py==1.11.0
pycryptodome==3.15.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32==304
regex==2022.3.2
requests==2.28.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toolz==0.12.0
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
yarl==1.8.1
zipp==3.9.0
